+++ packages/task/src/browser/task-service.ts
@@ -27,6 +27,7 @@ import { ProblemManager } from '@theia/markers/lib/browser/problem/problem-manag
 import { WorkspaceService } from '@theia/workspace/lib/browser/workspace-service';
 import { VariableResolverService } from '@theia/variable-resolver/lib/browser';
 import {
+    ContributedTaskConfiguration,
     ProblemMatcher,
     ProblemMatchData,
     ProblemMatcherContribution,
@@ -34,7 +35,6 @@ import {
     TaskExitedEvent,
     TaskInfo,
     TaskConfiguration,
-    TaskCustomization,
     TaskOutputProcessedEvent,
     RunTaskOption
 } from '../common';
@@ -203,13 +203,16 @@ export class TaskService implements TaskConfigurationClient {
 
     /** Returns an array of the task configurations configured in tasks.json and provided by the extensions. */
     async getTasks(): Promise<TaskConfiguration[]> {
-        const configuredTasks = this.getConfiguredTasks();
+        const configuredTasks = await this.getConfiguredTasks();
         const providedTasks = await this.getProvidedTasks();
-        return [...configuredTasks, ...providedTasks];
+        const notCustomizedProvidedTasks = providedTasks.filter(provided =>
+            !configuredTasks.some(configured => ContributedTaskConfiguration.equals(configured, provided))
+        );
+        return [...configuredTasks, ...notCustomizedProvidedTasks];
     }
 
     /** Returns an array of the task configurations which are configured in tasks.json files */
-    getConfiguredTasks(): TaskConfiguration[] {
+    getConfiguredTasks(): Promise<TaskConfiguration[]> {
         return this.taskConfigurations.getTasks();
     }
 
@@ -304,7 +307,7 @@ export class TaskService implements TaskConfigurationClient {
     async run(source: string, taskLabel: string): Promise<void> {
         let task = await this.getProvidedTask(source, taskLabel);
         const matchers: (string | ProblemMatcherContribution)[] = [];
-        if (!task) { // if a provided task cannot be found, search from tasks.json
+        if (!task) { // if a detected task cannot be found, search from tasks.json
             task = this.taskConfigurations.getTask(source, taskLabel);
             if (!task) {
                 this.logger.error(`Can't get task launch configuration for label: ${taskLabel}`);
@@ -317,9 +320,7 @@ export class TaskService implements TaskConfigurationClient {
                 }
             }
         } else { // if a provided task is found, check if it is customized in tasks.json
-            const taskType = task.taskType || task.type;
-            const customizations = this.taskConfigurations.getTaskCustomizations(taskType);
-            const matcherContributions = this.getProblemMatchers(task, customizations);
+            const matcherContributions = this.taskConfigurations.getProblemMatchers(task);
             matchers.push(...matcherContributions);
         }
         await this.problemMatcherRegistry.onReady();
@@ -381,27 +382,6 @@ export class TaskService implements TaskConfigurationClient {
         if (taskInfo.terminalId !== undefined) {
             this.attach(taskInfo.terminalId, taskInfo.taskId);
         }
-    }
-
-    private getProblemMatchers(taskConfiguration: TaskConfiguration, customizations: TaskCustomization[]): (string | ProblemMatcherContribution)[] {
-        const hasCustomization = customizations.length > 0;
-        const problemMatchers: (string | ProblemMatcherContribution)[] = [];
-        if (hasCustomization) {
-            const taskDefinition = this.taskDefinitionRegistry.getDefinition(taskConfiguration);
-            if (taskDefinition) {
-                const cus = customizations.filter(customization =>
-                    taskDefinition.properties.required.every(rp => customization[rp] === taskConfiguration[rp])
-                )[0]; // Only support having one customization per task
-                if (cus && cus.problemMatcher) {
-                    if (Array.isArray(cus.problemMatcher)) {
-                        problemMatchers.push(...cus.problemMatcher);
-                    } else {
-                        problemMatchers.push(cus.problemMatcher);
-                    }
-                }
-            }
-        }
-        return problemMatchers;
     }
 
     private async removeProblemMarks(option?: RunTaskOption): Promise<void> {